rag-assistant/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sentence-transformers==3.0.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.4
endee
groq
python-multipart
pymupdf
python-docx